feat: interfaces-b02-mcp-boundary-004 - Prove determinism, alias non-primacy, and behavior neutrality
