+++ pkg/mcp/factorysession/tool.go
@@ -4,6 +4,9 @@ package factorysession
 
 import (
 	"encoding/json"
+
+	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
+	"github.com/portpowered/infinite-you/pkg/apisurface"
 )
 
 // Tool names use Factory Session vocabulary and align with durable REST routes.
@@ -78,5 +81,75 @@ type ToolResponse[T any] struct {
 func (t ToolDefinition) MarshalJSON() ([]byte, error) {
 	type alias ToolDefinition
 	return json.Marshal(alias(t))
+}
+
+// DiscoverCompatibilityAliases returns workflow-named MCP tool aliases that
+// resolve to canonical Factory Session tool implementations.
+func DiscoverCompatibilityAliases() []CompatibilityAlias {
+	return []CompatibilityAlias{
+		{
+			Name:          ToolWorkflowValidate,
+			CanonicalName: ToolValidateSource,
+			Description: "Compatibility-only alias for you.factory_session.validate_source. " +
+				"Uses the same Factory preview validation contract and response shape.",
+			CompatibilityOnly: true,
+		},
+		{
+			Name:          ToolWorkflowRun,
+			CanonicalName: ToolStartSync,
+			Description: "Compatibility-only alias for you.factory_session.start_sync. " +
+				"Uses the same sync Factory Session start contract and response shape.",
+			CompatibilityOnly: true,
+		},
+		{
+			Name:          ToolWorkflowStatus,
+			CanonicalName: ToolGetSession,
+			Description: "Compatibility-only alias for you.factory_session.get. " +
+				"Uses the same durable Factory Session status read model.",
+			CompatibilityOnly: true,
+		},
+		{
+			Name:          ToolWorkflowResult,
+			CanonicalName: ToolGetResult,
+			Description: "Compatibility-only alias for you.factory_session.get_result. " +
+				"Uses the same durable Factory Session result read contract.",
+			CompatibilityOnly: true,
+		},
+		{
+			Name:          ToolWorkflowArtifacts,
+			CanonicalName: ToolListArtifacts,
+			Description: "Compatibility-only alias for you.factory_session.list_artifacts. " +
+				"Uses the same FactoryArtifact listing response shape.",
+			CompatibilityOnly: true,
+		},
+	}
+}
+
+// ResolveToolName maps one workflow compatibility alias to its canonical Factory
+// Session tool name. Unknown names pass through unchanged.
+func ResolveToolName(name string) string {
+	for _, alias := range DiscoverCompatibilityAliases() {
+		if alias.Name == name {
+			return alias.CanonicalName
+		}
+	}
+	return name
+}
+
+// ValidateSource runs the canonical Factory preview contract for the
+// you.factory_session.validate_source MCP tool without provider execution.
+func ValidateSource(input factoryapi.FactoryPreviewRequest) ToolResponse[factoryapi.FactoryPreviewResult] {
+	previewInput, err := apisurface.FactoryPreviewRequestFromAPI(input)
+	if err != nil {
+		envelope := requestValidationErrorEnvelope(err)
+		return ToolResponse[factoryapi.FactoryPreviewResult]{Error: &envelope}
+	}
+
+	preview := apisurface.FactoryPreviewResultFromPreview(apisurface.BuildFactoryPreview(previewInput))
+	if !preview.Valid {
+		envelope := validationErrorEnvelopeFromPreview(preview)
+		return ToolResponse[factoryapi.FactoryPreviewResult]{Error: &envelope}
+	}
+	return ToolResponse[factoryapi.FactoryPreviewResult]{Result: &preview}
 }
